requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
timm>=0.9.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
iterative-stratification>=0.1.7
Pillow>=10.0.0
tqdm>=4.65.0
